Added new constrollers for Estado, Cidade and it Service/Repository implementations

+++ src/main/java/com/algaworks/algafood/infrastructure/repository/EstadoRepositoryImpl.java
@@ -8,7 +8,9 @@ import javax.persistence.PersistenceContext;
 import com.algaworks.algafood.domain.model.Estado;
 import com.algaworks.algafood.domain.repository.EstadoRepository;
 
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 @Component
 public class EstadoRepositoryImpl implements EstadoRepository {
@@ -26,14 +28,21 @@ public class EstadoRepositoryImpl implements EstadoRepository {
         return entityManager.find(Estado.class, id);
     }
 
+    @Transactional
     @Override
     public Estado adicionar(Estado estado) {
         return entityManager.merge(estado);
     }
 
+    @Transactional
     @Override
-    public void remover(Estado estado) {
-        estado = buscar(estado.getId());
+    public void remover(Long estadoId) {
+        Estado estado = buscar(estadoId);
+        
+        if(estado == null ) {
+            throw new EmptyResultDataAccessException(1);
+        }
+
         entityManager.remove(estado);
     }
